+++ packages/react/src/atomic/Button/src/components/ButtonAnchor/buttonAnchor.types.ts
@@ -1,0 +1,9 @@
+import { AnchorHTMLAttributes } from "react";
+import { ButtonProps } from "../../button.types";
+
+type Extend = Omit<ButtonProps, "ButtonHTMLAttributes" | "disabled"> &
+  AnchorHTMLAttributes<HTMLAnchorElement>;
+
+export interface ButtonAnchorProps extends Extend {
+  href: string;
+}