+++ packages/react/src/atomic/Button/src/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./ButtonSkeleton";
+export * from "./ButtonAnchor";