+++ packages/react/src/atomic/Button/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The Button component allows us to initiate actions, make state or page changes.
 
+## 2023-02-06 `2.1.0`
+
+### 🎉 New features
+
+- Added new `Button.Anchor` subcomponent. [#78](https://github.com/TiendaNube/nimbus-design-system/pull/#78) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `Button.Anchor` stories documentation to `Button` stories. [#78](https://github.com/TiendaNube/nimbus-design-system/pull/#78) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2022-12-22 `2.0.0`
 
 ### 💡 Others